requirements.txt
openai
tiktoken
pdfplumber
langchain
azure-search-documents
pandas
openpyxl